pyproject.toml
[project]
name = "crud-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.119.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg[binary,pool]>=3.2.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
]
